microservices/admission_service/app/core/security.py
import os
from datetime import datetime, timezone
from pathlib import Path

from dotenv import load_dotenv
from jose import jwt, JWTError

BASE_DIR = Path(__file__).resolve().parent.parent.parent
load_dotenv(BASE_DIR / ".env_programme")

SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM", "HS256")

if not SECRET_KEY:
    raise RuntimeError("SECRET_KEY is not set. Check your .env_programme file.")


def decode_token(token: str) -> dict:
    return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
