microservices/admission_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
alembic
python-dotenv
pytest
httpx
python-jose[cryptography]
requests